@@ -9,11 +9,16 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@vitejs/plugin-vue2": "^2.3.1",
-    "vue": "^2.7.16",
-    "vuex": "^3.6.2"
+    "axios": "^1.7.3",
+    "pinia": "^2.2.1",
+    "vue": "^3.4.37",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
+    "@vitejs/plugin-vue": "^5.1.2",
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.4.41",
+    "tailwindcss": "^3.4.9",
     "vite": "^5.4.0"
   }
 }
